requirements.txt
pandas
numpy
PyYAML
rosbags
scipy
pandas
matplotlib
dateparser
paramiko
PyQt5
ping3
PyQt-Fluent-Widgets
pyyaml
pytz
pyvisa
pyserial
moviepy
Pillow
pexpect
openpyxl
reportlab
scp
wcwidth